tasks #13.0 Agregando Modal de Eliminar y Editar

+++ sistema_rh_frontend/src/empleados/ListadoEmpleado.jsx
@@ -1,7 +1,9 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
 import { NumericFormat } from "react-number-format";
-import { Link } from 'react-router-dom'
+import { Link } from 'react-router-dom';
+import Swal from "sweetalert2";
+import "sweetalert2/dist/sweetalert2.css";
 export default function ListadoEmpleado() {
 
   const URLBase = "http://localhost:8080/rh-app/empleados";
@@ -22,8 +24,24 @@ export default function ListadoEmpleado() {
   } 
   
   const eliminarEmpleado = async(id) => {
-    await axios.delete(`${URLBase}/${id}`)
+    Swal.fire({
+      title: "Estas seguro que deseas elimnar?",
+      text: "el registro  se eliminara permanente!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si, eliminar!",
+      cancelButtonText: "cancelar",
+    }).then(async (result) => {
+          if (result.isConfirmed) {
+            Swal.fire("Eliminado!", "Registro eliminado con exito", "success");
+            await axios.delete(`${URLBase}/${id}`)
+          }
     cargarEmpleados();
+
+    })
+    
   }
